replay: fix the Windows crash at startup and unthrottled playback
Two Windows-only failures in the replay library, which cabana shares:

- replay.exe segfaulted right after loading a route. currentSeconds()
  subtracts two uint64 timestamps, and at stream start cur_mono_time_
  is deliberately 1 ns behind route_start_ts_, so the difference wrapped
  to ~1.8e10 s. Casting that to int gives INT_MIN, the displayed route
  time went negative and the UCRT ctime() returned NULL where glibc
  prints a 1901 date; ConsoleUI then built a std::string from NULL.
  Make the difference signed (correct on every platform) and guard the
  ctime result.

- Playback ran unthrottled (cabana showed 1x but played ~20x). winpthreads'
  clock_nanosleep rejects CLOCK_MONOTONIC with EINVAL, which the loop
  treated as "done sleeping". Windows now waits on a high-resolution
  waitable timer in 5 ms slices, polling interrupt_requested between
  slices because there is no SIGUSR1 to cut the wait short. Measured:
  100 ms sleeps land within 0.5 ms, 2 ms sleeps within 0.3 ms, and a
  seek interrupts a 1 s sleep within ~7 ms.

+++ openpilot/tools/replay/replay.h
@@ -51,7 +51,8 @@ public:
   void setLoop(bool loop) { loop ? flags_ &= ~REPLAY_FLAG_NO_LOOP : flags_ |= REPLAY_FLAG_NO_LOOP; }
   bool loop() const { return !(flags_ & REPLAY_FLAG_NO_LOOP); }
   const Route &route() const { return seg_mgr_->route_; }
-  inline double currentSeconds() const { return double(cur_mono_time_ - route_start_ts_) / 1e9; }
+  // signed: cur_mono_time_ starts 1 ns before route_start_ts_ so the first event is not skipped
+  inline double currentSeconds() const { return double(int64_t(cur_mono_time_ - route_start_ts_)) / 1e9; }
   inline std::time_t routeDateTime() const { return route_date_time_; }
   inline uint64_t routeStartNanos() const { return route_start_ts_; }
   inline double toSeconds(uint64_t mono_time) const { return (mono_time - route_start_ts_) / 1e9; }
@@ -92,7 +93,7 @@ private:
   Timeline timeline_;
 
 #ifndef _WIN32
-  pthread_t stream_thread_id = 0;  // Windows: the stream thread sleep cannot be interrupted, it just runs out
+  pthread_t stream_thread_id = 0;  // Windows: precise_nano_sleep polls interrupt_requested_ instead of taking a signal
 #endif
   std::thread stream_thread_;
   std::mutex stream_lock_;
